+++ packages/styleguide/stories/Atoms/Button.stories.tsx
@@ -13,6 +13,7 @@ import {
   StoryTemplate,
   StoryStatus,
   StoryDescription,
+  decoratedStories,
   decoratedStory,
 } from '../Templating';
 
@@ -42,6 +43,8 @@ const deprecatedThemeKeys = [
   'purple',
 ];
 
+const statusThemeKeys = ['success', 'notice', 'error', 'announcement', 'info'];
+
 const themes = brandThemeKeys.reduce<Record<string, string>>(
   (previous, key) => ({ ...previous, [key]: key }),
   {}
@@ -52,13 +55,11 @@ const btnStyle = {
   marginBottom: '1rem',
 };
 
-const onClick = action('Clicked Button');
-
 const renderInlineButton = (theme: string, variant: string) => (
   <Button
     flat={variant === 'flat'}
     key={`${theme}`}
-    onClick={onClick}
+    onClick={action('Clicked Button')}
     outline={variant === 'outline'}
     style={btnStyle}
     theme={theme}
@@ -67,9 +68,9 @@ const renderInlineButton = (theme: string, variant: string) => (
   </Button>
 );
 
-export default decoratedStory('Atoms', Button);
+export default decoratedStories('Atoms', Button);
 
-export const allButtonThemes = () => {
+export const allButtonThemes = decoratedStory(() => {
   const variant = select('Variant', ['default', 'outline', 'flat'], 'default');
 
   return (
@@ -87,15 +88,20 @@ export const allButtonThemes = () => {
         {brandThemeKeys.map(theme => renderInlineButton(theme, variant))}
       </div>
       <br />
+      Alert status themes
+      <div>
+        {statusThemeKeys.map(theme => renderInlineButton(theme, variant))}
+      </div>
+      <br />
       We also have legacy button colors floating around. Do not use these.
       <div>
         {deprecatedThemeKeys.map(theme => renderInlineButton(theme, variant))}
       </div>
     </StoryTemplate>
   );
-};
+});
 
-export const buttonBaseVariants = () => (
+export const buttonBaseVariants = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.Ready}>
     <StoryDescription>
       Use <code>ButtonBase</code> when you need something that acts as a link or
@@ -119,9 +125,9 @@ export const buttonBaseVariants = () => (
       </ButtonBase>
     </div>
   </StoryTemplate>
-);
+));
 
-export const linkButton = () => (
+export const linkButton = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.Ready}>
     <p>
       This is an example of a{' '}
@@ -132,9 +138,9 @@ export const linkButton = () => (
       button, it will functionally behave as a link.
     </p>
   </StoryTemplate>
-);
+));
 
-export const standardButtonOptions = () => {
+export const standardButtonOptions = decoratedStory(() => {
   const round = boolean('Round', false);
 
   return (
@@ -151,7 +157,7 @@ export const standardButtonOptions = () => {
         </Button>
         <Button
           href="#"
-          onClick={onClick}
+          onClick={action('Clicked Button')}
           outline
           round={round}
           style={btnStyle}
@@ -164,7 +170,7 @@ export const standardButtonOptions = () => {
         </Button>
         <Button
           href="#"
-          onClick={onClick}
+          onClick={action('Clicked Button')}
           round={round}
           size="small"
           style={btnStyle}
@@ -189,9 +195,9 @@ export const standardButtonOptions = () => {
       </div>
     </StoryTemplate>
   );
-};
+});
 
-export const editablePlayground = () => (
+export const editablePlayground = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.Ready}>
     <StoryDescription>
       Play with this button, if you&apos;d like!
@@ -201,7 +207,7 @@ export const editablePlayground = () => (
       caps={boolean('caps', false)}
       go={boolean('go', false)}
       link={boolean('link', false)}
-      onClick={onClick}
+      onClick={action('Clicked Button')}
       outline={boolean('outline', false)}
       size={select('size', ['small', 'large', undefined], undefined)}
       theme={select('theme', themes, 'brand-blue')}
@@ -210,4 +216,4 @@ export const editablePlayground = () => (
       {text('Label', 'Submit')}
     </Button>
   </StoryTemplate>
-);
+));